- fixed warn: typecasting - fixed warn: empty arrays - fixed warn: no name struct - added name[]_size constant in bin2h - added some compile.cmd 'if exist' checks
* with -pedantic -fpermissive, mingw g++ seems 98% happy..
(some issues in the vstsdk 2.4)

+++ ax/axCanvasW32.h
@@ -228,7 +228,7 @@ class axCanvasImpl : public axCanvasBase
           float y1 = y + sin(a1*PI2) * size;
           float x2 = x + cos(a2*PI2) * size;
           float y2 = y + sin(a2*PI2) * size;
-          Arc(mDC,aX1,aY1,aX2,aY2,x1,y1,x2,y2);
+          Arc(mDC,aX1,aY1,aX2,aY2,(int)x1,(int)y1,(int)x2,(int)y2);
         }
       }
 
@@ -253,7 +253,7 @@ class axCanvasImpl : public axCanvasBase
           float y1 = y + sin(a1*PI2) * size;
           float x2 = x + cos(a2*PI2) * size;
           float y2 = y + sin(a2*PI2) * size;
-          Pie(mDC,aX1,aY1,aX2,aY2,x1,y1,x2,y2);
+          Pie(mDC,aX1,aY1,aX2,aY2,(int)x1,(int)y1,(int)x2,(int)y2);
         }
       }
